Rename tags/utils/utils.py to tags/utils.py and move remaining files from tags/utils to tags/admin
