+++ pontoon/tags/admin/__init__.py
@@ -1,0 +1,10 @@
+from .resources import TagsResourcesTool
+from .tags import TagsTool
+from .tag import TagTool
+
+
+__all__ = (
+    "TagsResourcesTool",
+    "TagsTool",
+    "TagTool",
+)
