+++ pontoon/tags/admin/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.utils.functional import cached_property
 
 from pontoon.base.models import Resource
-from pontoon.tags.utils import TagsTool
+from .tags import TagsTool
 
 
 class LinkTagResourcesAdminForm(forms.Form):
