--- pontoon/tags/utils/__init__.py
@@ -1,12 +1,0 @@
-from .resources import TagsResourcesTool
-from .utils import Tags
-from .tags import TagsTool
-from .tag import TagTool
-
-
-__all__ = (
-    "TagsResourcesTool",
-    "Tags",
-    "TagsTool",
-    "TagTool",
-)
